Fix compiler warnings.

+++ src/ias/io/reader/network/posix/posix_tcp_socket_reader.cpp
@@ -48,7 +48,7 @@ void PosixTcpSocketReader::closeReader( void ) {
 }
 
 std::size_t PosixTcpSocketReader::readByte( char * byte ) {
-    int nBytes;
+    long nBytes;
     
     nBytes = 0;
     if( mSocket->isConnected() ) {
@@ -64,7 +64,7 @@ std::size_t PosixTcpSocketReader::readByte( char * byte ) {
 
 std::size_t PosixTcpSocketReader::readBytes( char * buffer,
                                              const std::size_t bufferSize ) {
-    int nBytes;
+    long nBytes;
     
     // Checking the precondition.
     assert( buffer != nullptr && bufferSize > 0 );
